Move mapping classes into a separate package InferredMapperFactory -> InferredRowMapperFactory Move RegisteredRowMapper to a lambda on SqlStatement
