+++ core/src/main/java/org/jdbi/v3/MappingRegistry.java
@@ -22,6 +22,11 @@ import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
 
+import org.jdbi.v3.mapper.BuiltInMapperFactory;
+import org.jdbi.v3.mapper.ColumnMapperFactory;
+import org.jdbi.v3.mapper.InferredColumnMapperFactory;
+import org.jdbi.v3.mapper.InferredRowMapperFactory;
+import org.jdbi.v3.mapper.RowMapperFactory;
 import org.jdbi.v3.tweak.ColumnMapper;
 import org.jdbi.v3.tweak.RowMapper;
 import org.jdbi.v3.util.SingleColumnMapper;
@@ -51,7 +56,7 @@ class MappingRegistry
 
     public void addRowMapper(RowMapper<?> mapper)
     {
-        this.addRowMapper(new InferredMapperFactory(mapper));
+        this.addRowMapper(new InferredRowMapperFactory(mapper));
     }
 
     public void addRowMapper(RowMapperFactory factory)
