--- core/src/main/java/org/jdbi/v3/RegisteredRowMapper.java
@@ -1,41 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jdbi.v3;
-
-import java.lang.reflect.Type;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-import org.jdbi.v3.tweak.RowMapper;
-
-class RegisteredRowMapper<T> implements RowMapper<T>
-{
-
-    private final Type type;
-    private final MappingRegistry registry;
-
-    RegisteredRowMapper(Type type, MappingRegistry registry) {
-        this.type = type;
-        this.registry = registry;
-    }
-
-    @Override
-    @SuppressWarnings("unchecked")
-    public T map(ResultSet r, StatementContext ctx) throws SQLException
-    {
-        RowMapper<?> mapper = registry.findRowMapperFor(type, ctx)
-                .orElseThrow(() -> new UnsupportedOperationException("No mapper registered for " + type));
-        return (T) mapper.map(r, ctx);
-    }
-}
