+++ core/src/main/java/org/jdbi/v3/util/EnumByNameMapperFactory.java
@@ -18,8 +18,8 @@ import static org.jdbi.v3.Types.getErasedType;
 import java.lang.reflect.Type;
 import java.util.Optional;
 
-import org.jdbi.v3.ColumnMapperFactory;
 import org.jdbi.v3.StatementContext;
+import org.jdbi.v3.mapper.ColumnMapperFactory;
 import org.jdbi.v3.tweak.ColumnMapper;
 
 /**
