+++ core/src/test/java/org/jdbi/v3/BeanMapperTest.java
@@ -24,6 +24,7 @@ import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
+import org.jdbi.v3.mapper.BeanMapper;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
