+++ core/src/test/java/org/jdbi/v3/ValueTypeMapper.java
@@ -21,6 +21,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Optional;
 
+import org.jdbi.v3.mapper.ColumnMapperFactory;
 import org.jdbi.v3.tweak.ColumnMapper;
 
 public class ValueTypeMapper implements ColumnMapper<ValueType> {
