+++ postgres/src/main/java/org/jdbi/v3/pg/SqlArrayMapperFactory.java
@@ -16,9 +16,9 @@ package org.jdbi.v3.pg;
 import java.lang.reflect.Type;
 import java.util.Optional;
 
-import org.jdbi.v3.ColumnMapperFactory;
 import org.jdbi.v3.StatementContext;
 import org.jdbi.v3.Types;
+import org.jdbi.v3.mapper.ColumnMapperFactory;
 import org.jdbi.v3.tweak.ColumnMapper;
 
 public class SqlArrayMapperFactory implements ColumnMapperFactory {
