+++ sqlobject/src/main/java/org/jdbi/v3/sqlobject/customizers/RegisterColumnMapperFactory.java
@@ -21,7 +21,7 @@ import java.lang.annotation.Target;
 import java.lang.reflect.Method;
 
 import org.jdbi.v3.Query;
-import org.jdbi.v3.ColumnMapperFactory;
+import org.jdbi.v3.mapper.ColumnMapperFactory;
 import org.jdbi.v3.sqlobject.SqlStatementCustomizer;
 import org.jdbi.v3.sqlobject.SqlStatementCustomizerFactory;
 import org.jdbi.v3.sqlobject.SqlStatementCustomizingAnnotation;
